feat: Add InscripcionForm to CursoForm and create inscribir_alumno_curso view
This commit adds the InscripcionForm to the CursoForm in the administracion/forms.py file. It also creates the inscribir_alumno_curso view in the administracion/views.py file. The view allows authenticated users to submit an InscripcionForm to enroll in a Curso.

chore: Update Inscripcion model with evaluacion_docente field

This commit updates the Inscripcion model in the administracion/models.py file by adding the evaluacion_docente field. The field is a boolean field with a default value of False.

chore: Update administracion/urls.py

This commit adds a new URL path for the inscribir_alumno_curso view in the administracion/urls.py file.

chore: Update alumno_home.html template

This commit updates the alumno_home.html template in the autentificar/templates directory. It adds two links: one to the lista_cursos page and one to the evaluacion_alumno page.

feat: Add InscripcionForm and inscribir_alumno_curso view

This commit adds the InscripcionForm to the administracion/forms.py file and creates the inscribir_alumno_curso view in the administracion/views.py file. The view allows authenticated users to submit an InscripcionForm to enroll in a Curso.

feat: Add RespuestaForm to PreguntaForm

This commit adds the RespuestaForm to the PreguntaForm in the evaluacion/forms.py file. The RespuestaForm includes a required field for the respuesta.

chore: Update Respuesta model with RESPUESTAS_CHOICES

This commit updates the Respuesta model in the evaluacion/models.py file by adding the RESPUESTAS_CHOICES field. The field is an integer field with choices for different response options.

chore: Update evaluacion/urls.py

This commit adds a new URL path for the evaluacion_alumno view in the evaluacion/urls.py file.

feat: Add evaluacion_alumno view

This commit adds the evaluacion_alumno view to the evaluacion/views.py file. The view is accessible only to authenticated alumno users and displays a list of courses that the alumno is enrolled in and has not yet evaluated the docente for.

chore: Update evaluacion app

This commit includes various changes to the evaluacion app:
- Added evaluacion_alumno.html template
- Updated the evaluacion/urls.py file to include the evaluacion_alumno URL path
- Added the Inscripcion model from the administracion app to the imports in the evaluacion/views.py file
- Added the evaluacion_alumno view to the evaluacion/views.py file

feat: Add evaluacion_docente.html template

This commit adds the evaluacion_docente.html template to the evaluacion/templates directory.

chore: Add evaluacion, administracion apps and update urls

This commit adds the evaluacion and administracion apps to the project. It also updates the urls.py file in the saes/evaluacion and saes/administracion directories.

feat: Add evaluacion app with models, forms, and views

This commit adds the evaluacion app to the project. It includes the following changes:
- Added tests.py file to the evaluacion app
- Added admin.py file to the evaluacion app
- Added apps.py file to the evaluacion app
- Added models.py file to the evaluacion app, including the Pregunta and Respuesta models
- Added forms.py file to the evaluacion app, including the PreguntaForm and RespuestaForm forms
- Added urls.py file to the evaluacion app, including the create_pregunta, list_preguntas, evaluacion_docente, and evaluacion_alumno URLs
- Added templates for list_preguntas.html, create_pregunta.html, evaluacion_docente.html, and evaluacion_alumno.html